+++ clang/docs/ReleaseNotes.rst
@@ -210,6 +210,8 @@ AST Matchers
 clang-format
 ------------
 
+- Adds ``BreakBeforeTemplateCloser`` option.
+
 libclang
 --------